[SPIRV] Fix global hoisting issues
Hoisting relies on the duplicate tracker. The current implementation
hoists types first then constants. This will fail for `OpTypeArray`
given that it is a type and it has a constant operand.

Another problem was in `numberRegistersGlobally` function. We only add
dummy
registers if the target global register is null. When the target
register is not null and beyond the scope of the machine function
register lookup it fails. Now, we add dummy
registers after every mapping to a global register to resolve that.